id from long to int

+++ backend/src/main/java/ch/fhnw/ip6/ospp/vo/TeacherVO.java
@@ -3,7 +3,7 @@ package ch.fhnw.ip6.ospp.vo;
 
 public interface TeacherVO {
 
-    long getId();
+    int getId();
 
     String getFirstname();
